Fix bug when associative array is passed
The convert... methods will trigger whenever there are assetConverters available. This error triggers whevener JS files are specified with options; to take advantage of the attributes of the <script> tag. This change ensures that we use the correct URL when converting the file.

The use case which prompted this is import-maps. For import-maps, we would need to define JS script tags as modules and as import maps. With this change, script files can be defined independently ann injected with different types. A cleaner way to do this would be to have methods like "addJsModules()" and "addJsImportMaps()" to explicitly support this use case. This is a band-aid fix.